messages: Add suggested file name to the Attachment messages
In Cucumber JVM it is possible to include a file name when processing bytes:

```java
@When("I take a screenshot")
public void i_take_a_screen_shot() {
    byte[] bytes = ...
    scenario.attach(bytes, "image/png", "selfie.png");
}
```

Consumers can use this file name can be used as a suggestion when downloading
the file or when extracting the file from the report. This provides additional
clarity when many attachments are involved.

See: https://github.com/cucumber/cucumber-jvm/issues/2072